Deprecated Android tabs were replaced by Fragments. Skeleton of tab switching was written and save Parcel class was created. Type of latitude and longitude was changed to Long. Minimum android version is 3.1.

+++ PartyPoolEventsSearcherClient/src/cz/cvut/felk/via/data/Event.java
@@ -24,9 +24,9 @@ public class Event implements Serializable {
 
     private Date stopEvent;
 
-    private Double latitude;
+    private Long latitude;
 
-    private Double longitude;
+    private Long longitude;
 
     private List<CommentInfo> comments;
 
@@ -79,19 +79,19 @@ public class Event implements Serializable {
 		this.stopEvent = stopEvent;
 	}
 
-	public Double getLatitude() {
+	public Long getLatitude() {
 		return latitude;
 	}
 
-	public void setLatitude(Double latitude) {
+	public void setLatitude(Long latitude) {
 		this.latitude = latitude;
 	}
 
-	public Double getLongitude() {
+	public Long getLongitude() {
 		return longitude;
 	}
 
-	public void setLongitude(Double longitude) {
+	public void setLongitude(Long longitude) {
 		this.longitude = longitude;
 	}
